refactor(landing): inline About/Features/Pricing dans index.tsx
Seuls Hero et Servers restent des composants extraits (ils seront réutilisés
dans les pages /frame/* à venir). Les 3 autres sections étaient spécifiques
à la landing sans plan de réutilisation : on évite le surdécoupage en les
fondant dans la page (un seul index.tsx + index.module.css avec classes
préfixées par section).
