--- src/components/landing/About/styles.module.css
@@ -1,139 +1,0 @@
-.section {
-  overflow-x: clip;
-}
-
-.titleGrid {
-  display: grid;
-  grid-template-rows: auto;
-  grid-template-columns: 1.25fr 1fr;
-  grid-column-gap: 100px;
-  grid-row-gap: 100px;
-  margin-bottom: 30px;
-}
-
-.title {
-  font-family: var(--ifm-heading-font-family);
-  color: #fff;
-  font-size: 50px;
-  font-weight: 600;
-  line-height: 1.2em;
-  margin: 0;
-}
-
-.counterGrid {
-  display: grid;
-  grid-template-rows: auto;
-  grid-template-columns: 1fr 1fr;
-  grid-column-gap: 100px;
-  grid-row-gap: 100px;
-  align-items: center;
-}
-
-.imageWrap {
-  position: relative;
-}
-
-.curve {
-  width: 100%;
-  height: auto;
-  display: block;
-}
-
-.curveLine {
-  stroke-dasharray: 1400;
-  stroke-dashoffset: 1400;
-  animation: rpDraw 1.5s cubic-bezier(0.4, 0, 0.2, 1) forwards;
-}
-
-@keyframes rpDraw {
-  to {
-    stroke-dashoffset: 0;
-  }
-}
-
-.curvePulse {
-  transform-origin: 860px 28px;
-  animation: rpPulse 1.8s ease-out infinite;
-}
-
-@keyframes rpPulse {
-  0% {
-    transform: scale(1);
-    opacity: 0.9;
-  }
-  70% {
-    transform: scale(2.8);
-    opacity: 0;
-  }
-  100% {
-    transform: scale(2.8);
-    opacity: 0;
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .curveLine,
-  .curvePulse {
-    animation: none !important;
-  }
-  .curveLine {
-    stroke-dashoffset: 0;
-  }
-}
-
-.stats {
-  display: grid;
-  grid-template-rows: auto auto;
-  grid-template-columns: 1fr 1fr;
-  grid-column-gap: 60px;
-  grid-row-gap: 60px;
-  align-items: start;
-}
-
-.counterItem {
-  display: flex;
-  flex-flow: column;
-}
-
-.counterTitle {
-  font-family: var(--ifm-heading-font-family);
-  margin-bottom: 0;
-  font-size: 44px;
-  font-weight: 600;
-  line-height: 1.3em;
-  color: #fff;
-}
-
-.counterLabel {
-  color: #e1e0e9;
-  font-size: 18px;
-  margin-top: 4px;
-}
-
-@media screen and (max-width: 991px) {
-  .titleGrid {
-    grid-template-columns: 1fr;
-    grid-column-gap: 10px;
-    grid-row-gap: 10px;
-  }
-  .counterGrid {
-    grid-template-columns: 1fr;
-  }
-  .stats {
-    grid-column-gap: 50px;
-    grid-row-gap: 50px;
-  }
-}
-
-@media screen and (max-width: 767px) {
-  .title {
-    font-size: 42px;
-  }
-  .counterTitle {
-    font-size: 30px;
-  }
-  .stats {
-    grid-column-gap: 30px;
-    grid-row-gap: 30px;
-  }
-}
