--- src/components/landing/Features/styles.module.css
@@ -1,160 +1,0 @@
-.section {
-  position: relative;
-}
-
-.titleWrap {
-  z-index: 2;
-  position: relative;
-  margin: 0 auto 70px auto;
-  width: 50%;
-  text-align: center;
-  display: flex;
-  flex-flow: column;
-}
-
-.title {
-  margin: 0 0 10px 0;
-  font-family: var(--ifm-heading-font-family);
-  color: #fff;
-  font-size: 50px;
-  font-weight: 600;
-  line-height: 1.2em;
-}
-
-.subtitle {
-  margin: 0;
-  color: #e1e0e9;
-  font-size: 18px;
-  font-weight: 400;
-}
-
-.grid {
-  position: relative;
-  display: grid;
-  grid-template-rows: auto;
-  grid-template-columns: 1fr 1fr 1fr;
-  grid-column-gap: 50px;
-  grid-row-gap: 50px;
-}
-
-.decoration {
-  background-image: linear-gradient(180deg, #d35f5f 40%, #a561a3);
-  opacity: 0.5;
-  filter: blur(140px);
-  width: 450px;
-  height: 100px;
-  margin: auto;
-  position: absolute;
-  inset: 0;
-  z-index: 0;
-  pointer-events: none;
-}
-
-.item {
-  z-index: 2;
-  position: relative;
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-  text-align: center;
-  text-decoration: none;
-  transition: transform 0.3s;
-  column-gap: 20px;
-  row-gap: 20px;
-  color: inherit;
-}
-
-.item:hover {
-  transform: scale(1.04);
-  text-decoration: none;
-  color: inherit;
-}
-
-.iconWrap {
-  flex: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 80px;
-  height: 80px;
-  padding: 1px;
-  border-radius: 50%;
-  background-image: linear-gradient(
-    135deg,
-    #726cb3,
-    #d35f5f 60%,
-    #a561a3
-  );
-}
-
-.iconBg {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-  background-color: #1b1a25;
-}
-
-.icon {
-  width: 32px;
-  height: 32px;
-}
-
-.itemTitle {
-  margin: 0 0 10px 0;
-  font-family: var(--ifm-heading-font-family);
-  color: #fff;
-  font-size: 26px;
-  font-weight: 600;
-  line-height: 1.3em;
-}
-
-.itemDescription {
-  margin: 0;
-  color: #e1e0e9;
-  font-size: 18px;
-}
-
-@media screen and (max-width: 1279px) {
-  .grid {
-    grid-column-gap: 50px;
-    grid-row-gap: 50px;
-  }
-}
-
-@media screen and (min-width: 1280px) {
-  .grid {
-    grid-column-gap: 90px;
-    grid-row-gap: 100px;
-  }
-  .titleWrap {
-    width: 95%;
-  }
-}
-
-@media screen and (max-width: 991px) {
-  .titleWrap {
-    width: 100%;
-  }
-  .grid {
-    grid-template-columns: 1fr 1fr;
-  }
-  .iconWrap {
-    width: 70px;
-    height: 70px;
-  }
-}
-
-@media screen and (max-width: 767px) {
-  .title {
-    font-size: 42px;
-  }
-  .itemTitle {
-    font-size: 22px;
-  }
-  .grid {
-    grid-column-gap: 30px;
-  }
-}
