--- src/components/landing/Pricing/styles.module.css
@@ -1,202 +1,0 @@
-.section {
-  overflow-x: clip;
-  padding-top: 150px;
-  padding-bottom: 150px;
-}
-
-.titleWrap {
-  text-align: center;
-  width: 60%;
-  margin: 0 auto 40px auto;
-  position: relative;
-}
-
-.titleInner {
-  z-index: 2;
-  position: relative;
-}
-
-.title {
-  margin: 0 0 10px 0;
-  font-family: var(--ifm-heading-font-family);
-  color: #fff;
-  font-size: 50px;
-  font-weight: 600;
-  line-height: 1.2em;
-}
-
-.description {
-  margin: 0 0 40px 0;
-  color: #e1e0e9;
-  font-size: 18px;
-}
-
-.decoration {
-  background-image: linear-gradient(180deg, #d35f5f 40%, #a561a3);
-  filter: blur(110px);
-  width: 50%;
-  height: 90px;
-  margin: 179px auto auto auto;
-  position: absolute;
-  inset: 0;
-  z-index: 0;
-  pointer-events: none;
-}
-
-.grid {
-  z-index: 2;
-  position: relative;
-  display: grid;
-  grid-template-rows: auto;
-  grid-template-columns: 1fr 1fr 1fr;
-  grid-column-gap: 30px;
-  grid-row-gap: 30px;
-  margin-bottom: 30px;
-}
-
-.item {
-  display: flex;
-  flex-flow: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  border: 1px solid #242328;
-  border-radius: 20px;
-  background-color: #070417;
-  padding: 30px;
-}
-
-.itemCenter {
-  border-color: #d35f5f;
-  background-color: #1b1a25;
-}
-
-.preTitle {
-  border: 1px solid #d35f5f;
-  border-radius: 50px;
-  background-color: #242328;
-  color: #fff;
-  text-align: center;
-  margin-bottom: 20px;
-  padding: 8px 20px;
-  font-size: 16px;
-  line-height: 1.3em;
-  font-family: var(--ifm-font-family-base);
-}
-
-.preTitleFounder {
-  border-color: #ffab22;
-  background-color: #1b1a25;
-}
-
-.preTitleEnterprise {
-  border-color: #5f6fd3;
-}
-
-.priceWrap {
-  display: flex;
-  flex-flow: wrap;
-  justify-content: flex-start;
-  align-items: baseline;
-  width: 100%;
-  column-gap: 10px;
-  row-gap: 5px;
-  margin-bottom: 20px;
-}
-
-.price {
-  margin: 0;
-  font-family: var(--ifm-heading-font-family);
-  color: #fff;
-  font-size: 30px;
-  font-weight: 600;
-  line-height: 1.2em;
-}
-
-.priceCenter {
-  text-align: right;
-  background-image: linear-gradient(180deg, #d35f5f 40%, #a561a3);
-  -webkit-text-fill-color: transparent;
-  -webkit-background-clip: text;
-  background-clip: text;
-}
-
-.itemTagline {
-  margin: 0;
-  color: #e1e0e9;
-  font-size: 18px;
-  line-height: 1.5em;
-}
-
-.featureList {
-  display: flex;
-  flex-flow: column;
-  width: 100%;
-  margin-top: 50px;
-  margin-bottom: 50px;
-  column-gap: 15px;
-  row-gap: 15px;
-}
-
-.featureItem {
-  display: flex;
-  align-items: flex-start;
-  justify-content: flex-start;
-  column-gap: 8px;
-  row-gap: 8px;
-  color: #e1e0e9;
-}
-
-.featureIcon {
-  flex: none;
-  width: 18px;
-  height: 18px;
-  margin-top: 5px;
-}
-
-.buttonList {
-  display: flex;
-  flex-flow: column;
-  align-items: stretch;
-  justify-content: flex-start;
-  width: 100%;
-  margin-top: auto;
-  text-align: center;
-  column-gap: 15px;
-  row-gap: 15px;
-}
-
-@media screen and (max-width: 991px) {
-  .section {
-    padding-top: 100px;
-    padding-bottom: 100px;
-  }
-  .titleWrap {
-    width: 100%;
-  }
-  .grid {
-    grid-template-columns: 1fr 1fr;
-  }
-  .featureList {
-    margin-top: 30px;
-    margin-bottom: 30px;
-  }
-}
-
-@media screen and (max-width: 767px) {
-  .section {
-    padding-top: 80px;
-    padding-bottom: 80px;
-  }
-  .title {
-    font-size: 42px;
-  }
-  .grid {
-    grid-template-columns: 1fr;
-    grid-column-gap: 20px;
-    grid-row-gap: 20px;
-    margin-bottom: 20px;
-  }
-  .item {
-    padding: 20px;
-  }
-}
